core/vm: drop a redundant uint64 conversion
evm.Context.Time is already uint64, and the multiplier-ceiling test wrapped it
anyway. golangci-lint's unconvert catches it; the local run that preceded this
test did not, because the test came after it.

+++ core/vm/b20_asset_test.go
@@ -480,7 +480,7 @@ func TestB20MultiplierCeilingIsOneConstant(t *testing.T) {
 	}
 	ceiling := new(uint256.Int).SetBytes(advertised)
 	over := new(uint256.Int).AddUint64(ceiling, 1)
-	future := u256hash(uint64(evm.Context.Time) + 3600)
+	future := u256hash(evm.Context.Time + 3600)
 
 	for _, tc := range []struct {
 		name  string
